requirements.txt
certifi==2020.6.20
cycler==0.10.0
decorator==4.4.2
future==0.18.2
h5py==2.10.0
imageio==2.9.0
kiwisolver==1.2.0
matplotlib==3.3.2
networkx==2.5
numpy==1.19.0
opencv-python==4.2.0.34
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
scikit-image==0.17.2
scipy==1.5.2
six==1.15.0
tifffile==2020.10.1
torch==1.5.1
torchvision==0.6.1
